Refactor compare_files() and set instead of return
The DataFrames created in compare_files() will be useful
in other parts of the code. Particularly, for creating
a table similar to the exact_differences_table
and for implementing validation criteria for factors.
Break calculating file number of differences in a new
method. Also, set DataFrames as properties rather than return values
to make the DataFrames more accessible.